feat(entities/repo): Added models/entities for recipe and recipeItem - Added repo/interface for recipes

+++ src/main/java/org/example/dto/RecipeResponse.java
@@ -3,7 +3,7 @@ package org.example.dto;
 import java.util.List;
 
 public record RecipeResponse (
-    long id,
+    Integer id,
     String title,
     String instructions,
     List<RecipeItemResponse> items
